#76: refactor POST unit and e2e tests
